+++ tests/Unit/Authentication/NativeAuthenticatorTest.php
@@ -19,46 +19,29 @@ declare(strict_types=1);
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-namespace Youthweb\Api\Tests\Unit;
+namespace Youthweb\Api\Tests\Unit\Authentication;
 
-use Youthweb\Api\YouthwebAuthenticator;
+use League\OAuth2\Client\Provider\AbstractProvider;
+use League\OAuth2\Client\Token\AccessTokenInterface;
+use PHPUnit\Framework\TestCase;
+use Youthweb\Api\Authentication\NativeAuthenticator;
 
-class YouthwebAuthenticatorTest extends \PHPUnit\Framework\TestCase
+class YouthwebAuthenticatorTest extends TestCase
 {
-    /**
-     * Create a Authenticator with mocks of all collaborators
-     */
-    private function createAuthenticator(array $options = [], array $collaborators = [])
-    {
-        $default_options = [];
-
-        $options = array_merge($default_options, $options);
-
-        $default_collaborators = [
-            'oauth2_provider' => $this->createMock('Youthweb\OAuth2\Client\Provider\Youthweb'),
-        ];
-
-        $collaborators = array_merge($default_collaborators, $collaborators);
-
-        return new YouthwebAuthenticator($options, $collaborators);
-    }
-
     /**
      * @test
      */
     public function testGetAuthorizationUrlReturnsUrl(): void
     {
-        $oauth2_provider = $this->createMock('Youthweb\OAuth2\Client\Provider\Youthweb');
+        $oauth2Provider = $this->createMock(AbstractProvider::class);
 
         $url = 'https://example.org';
 
-        $oauth2_provider->expects($this->once())
+        $oauth2Provider->expects($this->once())
             ->method('getAuthorizationUrl')
             ->willReturn($url);
 
-        $authenticator = $this->createAuthenticator([], [
-            'oauth2_provider' => $oauth2_provider,
-        ]);
+        $authenticator = new NativeAuthenticator($oauth2Provider);
 
         $this->assertSame($url, $authenticator->getAuthorizationUrl());
     }
@@ -68,17 +51,15 @@ class YouthwebAuthenticatorTest extends \PHPUnit\Framework\TestCase
      */
     public function testGetStateReturnsState(): void
     {
-        $oauth2_provider = $this->createMock('Youthweb\OAuth2\Client\Provider\Youthweb');
+        $oauth2Provider = $this->createMock(AbstractProvider::class);
 
         $state = 'random_string';
 
-        $oauth2_provider->expects($this->once())
+        $oauth2Provider->expects($this->once())
             ->method('getState')
             ->willReturn($state);
 
-        $authenticator = $this->createAuthenticator([], [
-            'oauth2_provider' => $oauth2_provider,
-        ]);
+        $authenticator = new NativeAuthenticator($oauth2Provider);
 
         $this->assertSame($state, $authenticator->getState());
     }
@@ -88,17 +69,19 @@ class YouthwebAuthenticatorTest extends \PHPUnit\Framework\TestCase
      */
     public function testGetStateWorkaroundReturnsState(): void
     {
-        $oauth2_provider = $this->createMock('Youthweb\OAuth2\Client\Provider\Youthweb');
+        $oauth2Provider = $this->createMock(AbstractProvider::class);
+
+        $oauth2Provider->expects($this->once())
+            ->method('getAuthorizationUrl')
+            ->willReturn('');
 
         $state = 'random_string';
 
-        $oauth2_provider->expects($this->exactly(2))
+        $oauth2Provider->expects($this->exactly(2))
             ->method('getState')
             ->will($this->onConsecutiveCalls('', $state));
 
-        $authenticator = $this->createAuthenticator([], [
-            'oauth2_provider' => $oauth2_provider,
-        ]);
+        $authenticator = new NativeAuthenticator($oauth2Provider);
 
         $this->assertSame($state, $authenticator->getState());
     }
@@ -108,7 +91,7 @@ class YouthwebAuthenticatorTest extends \PHPUnit\Framework\TestCase
      */
     public function testGetAccessTokenWithWrongGrantThrowsException(): void
     {
-        $authenticator = $this->createAuthenticator();
+        $authenticator = new NativeAuthenticator($this->createMock(AbstractProvider::class));
 
         $this->expectException('InvalidArgumentException');
         $this->expectExceptionMessage('Unsupported grant "wrong_grant"');
@@ -121,26 +104,17 @@ class YouthwebAuthenticatorTest extends \PHPUnit\Framework\TestCase
      */
     public function testGetAccessTokenWithAuthCodeAndStateSavesToken(): void
     {
-        $oauth2_provider = $this->createMock('Youthweb\OAuth2\Client\Provider\Youthweb');
-        $access_token = $this->createMock('League\OAuth2\Client\Token\AccessToken');
+        $oauth2Provider = $this->createMock(AbstractProvider::class);
+        $accessToken = $this->createMock(AccessTokenInterface::class);
 
-        $oauth2_provider->expects($this->once())
+        $oauth2Provider->expects($this->once())
             ->method('getAccessToken')
-            ->willReturn($access_token);
+            ->willReturn($accessToken);
 
-        $authenticator = $this->createAuthenticator(
-            [
-                'client_id'     => 'client_id',
-                'client_secret' => 'client_secret',
-                'redirect_url'  => 'https://example.org/callback',
-            ],
-            [
-                'oauth2_provider' => $oauth2_provider,
-            ]
-        );
+        $authenticator = new NativeAuthenticator($oauth2Provider);
 
         $this->assertSame(
-            $access_token,
+            $accessToken,
             $authenticator->getAccessToken('authorization_code', [
                 'code' => 'auth_code',
                 'state' => 'random_string',